src/main/java/com/lauradebella/urlshortener/Url/UrlController.java
package com.lauradebella.urlshortener.Url;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;

@RestController
public class UrlController {

    private UrlService urlService;

    @Autowired
    public UrlController(UrlService urlService) {
        this.urlService = urlService;
    }

    @RequestMapping(value = "/short",  method= RequestMethod.POST, produces = APPLICATION_JSON_UTF8_VALUE)
    public Url shorten(@Valid @RequestBody UrlRequest params) {
        return urlService.shortUrl(params.getLongUrl());
    }

}
