src/test/java/com/lauradebella/urlshortener/Url/UrlControllerTest.java
package com.lauradebella.urlshortener.Url;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.BDDMockito.given;
import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(UrlController.class)
public class UrlControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private ObjectMapper objectMapper;

    @MockBean
    private UrlService urlService;

    @Test
    public void shorten() throws Exception {
        UrlRequest request = new UrlRequest("http://www.longurl.com");
        Url response = Url.builder()
                .longUrl("http://www.longurl.com")
                .shortUrl("localhost/abc")
                .shortUrlId("abc")
                .build();

        given(this.urlService.shortUrl("http://www.longurl.com")).willReturn(response);
        mockMvc.perform(post("/short").contentType(APPLICATION_JSON_UTF8_VALUE)
                .content(objectMapper.writeValueAsString(request)))
                .andExpect(status().isOk())
                .andExpect(content().contentType(APPLICATION_JSON_UTF8_VALUE))
                .andExpect(jsonPath("$.longUrl").value("http://www.longurl.com"))
                .andExpect(jsonPath("$.shortUrl").value("localhost/abc"));
    }
}